Added security for endpoints

+++ simple-accounting/src/main/java/accounting/controller/AccountController.java
@@ -2,8 +2,12 @@ package accounting.controller;
 
 import java.util.List;
 
+import org.apache.catalina.User;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -69,6 +73,9 @@ public class AccountController {
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> deleteAccount(@PathVariable String id) {
+		if (!isUserAuthorized()) {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
 		accountService.deleteAccount(id);
 		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
@@ -98,6 +105,18 @@ public class AccountController {
 			@RequestParam("maxBalance") double maxBalance) {
 		List<AccountDTO> accounts = accountService.getAccountByAmountRange(minBalance, maxBalance);
 		return new ResponseEntity<>(accounts, HttpStatus.OK);
+	}
+
+	private boolean isUserAuthorized() {
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		if (authentication != null && authentication.isAuthenticated()) {
+			Object principal = authentication.getPrincipal();
+			if (principal instanceof UserDetails) {
+				UserDetails userDetails = (UserDetails) principal;
+				return userDetails.getAuthorities().stream()
+						.anyMatch(authory -> authory.getAuthority().equals("ROLE_ADMIN"));
+			}
+		}
 	}
 
 }
